kmod: import the entire hw code from the lib except for static config tables
* And make minor (at this point) adjustments so that it integrates a little
  better with the kernel coding style.
