+++ src/kmod/sja1105-reset.c
@@ -1,43 +1,17 @@
-/******************************************************************************
- * Copyright (c) 2016, NXP Semiconductors
- * All rights reserved.
+/*
+ * SPDX-License-Identifier: (GPL-2.0 OR BSD-3-Clause)
  *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice,
- * this list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- *
- * 3. Neither the name of the copyright holder nor the names of its
- * contributors may be used to endorse or promote products derived from this
- * software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- *****************************************************************************/
+ * Copyright (c) 2016-2018, NXP Semiconductors
+ */
 #include <lib/include/static-config.h>
 #include <lib/include/gtable.h>
 #include <lib/include/spi.h>
 #include <common.h>
+#include "sja1105.h"
 
 static void
-sja1105_reset_cmd_access(void *buf,
-                          struct sja1105_reset_cmd *reset,
-                          int write,
-                          uint64_t device_id)
+sja1105_reset_cmd_access(void *buf, struct sja1105_reset_cmd *reset,
+                         int write, uint64_t device_id)
 {
 	int  (*pack_or_unpack)(void*, uint64_t*, int, int, int);
 	int    size = 4;
@@ -62,45 +36,28 @@ sja1105_reset_cmd_access(void *buf,
 		pack_or_unpack(buf, &reset->por_rst,    1, 1, 4);
 	}
 }
-
-void sja1105_reset_cmd_pack(void *buf,
-                             struct sja1105_reset_cmd *reset,
-                             uint64_t device_id)
-{
-	sja1105_reset_cmd_access(buf, reset, 1, device_id);
-}
-
-void sja1105_reset_cmd_unpack(void *buf,
-                               struct sja1105_reset_cmd *reset,
-                               uint64_t device_id)
-{
-	sja1105_reset_cmd_access(buf, reset, 0, device_id);
-}
+#define sja1105_reset_cmd_pack(buf, reset, device_id) \
+	sja1105_reset_cmd_access(buf, reset, 1, device_id)
+#define sja1105_reset_cmd_unpack(buf, reset, device_id) \
+	sja1105_reset_cmd_access(buf, reset, 0, device_id)
 
 void sja1105_reset_cmd_show(struct sja1105_reset_cmd *reset)
 {
-	if (reset->switch_rst) {
+	if (reset->switch_rst)
 		logv("Main reset for all functional modules requested");
-	}
-	if (reset->cfg_rst) {
+	if (reset->cfg_rst)
 		logv("Chip configuration reset requested");
-	}
-	if (reset->car_rst) {
+	if (reset->car_rst)
 		logv("Clock and reset control logic reset requested");
-	}
-	if (reset->otp_rst) {
+	if (reset->otp_rst)
 		logv("OTP read cycle for reading product "
 		     "config settings requested");
-	}
-	if (reset->warm_rst) {
+	if (reset->warm_rst)
 		logv("Warm reset requested");
-	}
-	if (reset->cold_rst) {
+	if (reset->cold_rst)
 		logv("Cold reset requested");
-	}
-	if (reset->por_rst) {
+	if (reset->por_rst)
 		logv("Power-on reset requested");
-	}
 }
 
 int sja1105_reset_cmd_commit(struct sja1105_spi_setup *spi_setup,
@@ -123,11 +80,8 @@ int sja1105_reset_cmd_commit(struct sja1105_spi_setup *spi_setup,
 	}
 	sja1105_reset_cmd_pack(packed_buf, reset, spi_setup->device_id);
 
-	rc = sja1105_spi_send_packed_buf(spi_setup,
-	                                 SPI_WRITE,
-	                                 RGU_ADDR,
-	                                 packed_buf,
-	                                 BUF_LEN);
+	rc = sja1105_spi_send_packed_buf(spi_setup, SPI_WRITE, RGU_ADDR,
+	                                 packed_buf, BUF_LEN);
 out:
 	return rc;
 }
