--- src/lib/clock/sgmii.c
@@ -1,52 +1,0 @@
-/******************************************************************************
- * Copyright (c) 2018, NXP Semiconductors
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice,
- * this list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- *
- * 3. Neither the name of the copyright holder nor the names of its
- * contributors may be used to endorse or promote products derived from this
- * software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- *****************************************************************************/
-#include <lib/include/gtable.h>
-#include <lib/include/clock.h>
-#include <lib/include/spi.h>
-#include <common.h>
-
-/* TODO:
- * Standard clause 22 registers for the internal SGMII PCS are
- * memory-mapped starting at SPI address 0x1F0000.
- * The SGMII port should already have a basic initialization done
- * through the static configuration tables.
- * If any further SGMII initialization steps (autonegotiation or
- * checking the link training status) need to be done, they
- * might as well be added here.
- */
-int sgmii_clocking_setup(struct sja1105_spi_setup __attribute__((unused)) *spi_setup,
-                         int port, int speed_mbps)
-{
-	logv("TODO: Configure SGMII clocking for port %d speed %dMbps.",
-	     port, speed_mbps);
-	return 0;
-}
-
