+++ src/lib/static-config/static-config.c
@@ -30,7 +30,6 @@
  *****************************************************************************/
 #include <lib/include/static-config.h>
 #include <lib/include/gtable.h>
-#include <lib/include/port-control.h>
 #include <common.h>
 
 static void sja1105_table_write_crc(char *table_start, char *crc_ptr)
@@ -666,132 +665,49 @@ sja1105_static_config_get_length(struct sja1105_static_config *config)
 	return sum;
 }
 
-static int
-static_config_upload(struct sja1105_spi_setup *spi_setup,
-                     struct sja1105_static_config *config)
+const char *SJA1105E_DEVICE_ID_STR        = "SJA1105E";
+const char *SJA1105T_DEVICE_ID_STR        = "SJA1105T";
+const char *SJA1105P_DEVICE_ID_STR        = "SJA1105P";
+const char *SJA1105Q_DEVICE_ID_STR        = "SJA1105Q";
+const char *SJA1105R_DEVICE_ID_STR        = "SJA1105R";
+const char *SJA1105S_DEVICE_ID_STR        = "SJA1105S";
+const char *SJA1105PR_DEVICE_ID_STR       = "SJA1105P or SJA1105R";
+const char *SJA1105QS_DEVICE_ID_STR       = "SJA1105Q or SJA1105S";
+const char *SJA1105_NO_DEVICE_ID_STR      = "None";
+
+const char *sja1105_device_id_string_get(uint64_t device_id, uint64_t part_nr)
 {
-	struct   sja1105_table_header final_header;
-	char    *final_header_ptr;
-	char  *config_buf;
-	int    config_buf_len;
-	int    crc_len;
-	int    rc;
-
-	config_buf_len = sja1105_static_config_get_length(config);
-	config_buf = (char*) calloc(1, config_buf_len * sizeof(char));
-	if (!config_buf) {
-		loge("malloc failed");
-		rc = -ENOMEM;
-		goto out;
+	if (device_id == SJA1105E_DEVICE_ID) {
+		return SJA1105E_DEVICE_ID_STR;
 	}
-	/* Write Device ID and config tables to config_buf */
-	rc = sja1105_static_config_pack(config_buf, config);
-	if (rc < 0) {
-		loge("sja1105_static_config_pack failed");
-		goto out_free;
+	if (device_id == SJA1105T_DEVICE_ID) {
+		return SJA1105T_DEVICE_ID_STR;
 	}
-	/* Recalculate CRC of the last header */
-	/* Don't include the CRC field itself */
-	crc_len = config_buf_len - 4;
-	/* Read the whole table header */
-	final_header_ptr = config_buf + config_buf_len - SIZE_TABLE_HEADER;
-	sja1105_table_header_unpack(final_header_ptr, &final_header);
-	/* Modify */
-	final_header.crc = ether_crc32_le(config_buf, crc_len);
-	/* Rewrite */
-	sja1105_table_header_pack(final_header_ptr, &final_header);
-
-	rc = sja1105_spi_send_long_packed_buf(spi_setup,
-	                                      SPI_WRITE,
-	                                      CONFIG_ADDR,
-	                                      config_buf,
-	                                      config_buf_len);
-out_free:
-	free(config_buf);
-out:
-	return rc;
-}
-
-int sja1105_static_config_flush(struct sja1105_spi_setup *spi_setup,
-                                struct sja1105_static_config *config)
-{
-	struct sja1105_general_status status;
-	struct sja1105_egress_port_mask port_mask;
-	int i, rc;
-
-	rc = sja1105_static_config_check_valid(config);
-	if (rc < 0) {
-		loge("cannot upload config, because it is not valid");
-		goto staging_area_invalid_error;
-	}
-	/* Workaround for PHY jabbering during switch reset */
-	memset(&port_mask, 0, sizeof(port_mask));
-	for (i = 0; i < SJA1105T_NUM_PORTS; i++) {
-		port_mask.inhibit_tx[i] = 1;
-	}
-	rc = sja1105_inhibit_tx(spi_setup, &port_mask);
-	if (rc < 0) {
-		loge("sja1105_set_egress_port_mask failed");
-		goto hardware_not_responding_error;
-	}
-	/* Wait for an eventual egress packet to finish transmission
-	 * (reach IFG). It is guaranteed that a second one will not
-	 * follow, and that switch cold reset is thus safe
+	/* P and R have same Device ID, and differ by Part Number.
+	 * Same do Q and S.
 	 */
-	usleep(1000);
-	/* Put the SJA1105 in programming mode */
-	rc = sja1105_cold_reset(spi_setup);
-	if (rc < 0) {
-		loge("sja1105_reset failed");
-		goto hardware_left_floating_error;
+	if (IS_P(device_id, part_nr)) {
+		return SJA1105P_DEVICE_ID_STR;
 	}
-	rc = static_config_upload(spi_setup, config);
-	if (rc < 0) {
-		loge("static_config_upload failed");
-		goto hardware_left_floating_error;
+	if (IS_Q(device_id, part_nr)) {
+		return SJA1105Q_DEVICE_ID_STR;
 	}
-	/* Configure the CGU (PHY link modes and speeds) */
-	rc = sja1105_clocking_setup(spi_setup, &config->xmii_params[0],
-	                           &config->mac_config[0]);
-	if (rc < 0) {
-		loge("sja1105_clocking_setup failed");
-		goto hardware_left_floating_error;
+	if (IS_R(device_id, part_nr)) {
+		return SJA1105P_DEVICE_ID_STR;
 	}
-	/* Check that SJA1105 responded well to the config upload */
-	if (spi_setup->dry_run == 0) {
-		/* These checks simply cannot pass (and do not even
-		 * make sense to have) if we are in dry run mode */
-		rc = sja1105_general_status_get(spi_setup, &status);
-		if (rc < 0) {
-			goto hardware_left_floating_error;
-		}
-		if (status.ids == 1) {
-			loge("Mismatch between hardware and staging area "
-			     "device id. Wrote 0x%" PRIx64 ", wants 0x%" PRIx64,
-			     config->device_id, spi_setup->device_id);
-			goto hardware_left_floating_error;
-		}
-		if (status.crcchkl == 1) {
-			loge("local crc failed while uploading config");
-			goto hardware_left_floating_error;
-		}
-		if (status.crcchkg == 1) {
-			loge("global crc failed while uploading config");
-			goto hardware_left_floating_error;
-		}
-		if (status.configs == 0) {
-			loge("configuration is invalid");
-			goto hardware_left_floating_error;
-		}
+	if (IS_S(device_id, part_nr)) {
+		return SJA1105S_DEVICE_ID_STR;
 	}
-	return SJA1105_ERR_OK;
-staging_area_invalid_error:
-	sja1105_err_remap(rc, SJA1105_ERR_STAGING_AREA_INVALID);
-	return rc;
-hardware_left_floating_error:
-	sja1105_err_remap(rc, SJA1105_ERR_UPLOAD_FAILED_HW_LEFT_FLOATING);
-	return rc;
-hardware_not_responding_error:
-	sja1105_err_remap(rc, SJA1105_ERR_HW_NOT_RESPONDING);
-	return rc;
+	/* Fallback: if we don't know/care what the part_nr is, and we
+	 * have a P/R, we can simply pass -1 to part_nr and have this
+	 * function say it's either P or R, instead of reporting it
+	 * as invalid.
+	 */
+	if (device_id == SJA1105PR_DEVICE_ID) {
+		return SJA1105PR_DEVICE_ID_STR;
+	}
+	if (device_id == SJA1105QS_DEVICE_ID) {
+		return SJA1105QS_DEVICE_ID_STR;
+	}
+	return SJA1105_NO_DEVICE_ID_STR;
 }
